Send config bus into layers.v module
Configuration bus is a broadcast bus that sends all configuration data to
modules that can then accept the data if its addressed to them. The layers
module gets 4 different configuration values packed into one 32 bit cfg_data.

Configuration but address parameters are being stored into a verilog header
file that gets included by the module if it is to decode the config bus.