Add bias_add module to the layers module
Create a new input bus into the layers module that contains the convolutional
bias. The width of the bias_bus is the same as the kernel_bus as its being read
from the same memory in the kernel_mem module. However, instead of the bias_bus
containing the same number of values as there are defined by the GROUP_NB
parameter, the bias_bus will contain only one value per coulomb. The width and
format of the bias number/value is already the same as that of the intermediate
number values, thus no manipulation is needed before use.