Store bias values with kernel data in kernel_mem
The first address of the kernel read section will now be reserved for bias
values. When the read section is set to a new position the first address is
read so that the bias values associated with the convolution will be presented
to the layers.v module. When the read pointer wraps around the read section the
bias address is skipped and only the kernel values are read again.