Add image_mem module and testbench
The image_mem module has a single port memory that stores the image data for
convolutions. If the wr data stream is wider then the read, the memory is
separated into coulombs so that the write words can be written as one and the
reading can access the data per coulomb.