Add skeleton image_read module and testbench
The image_read module generates the read requests for data from the image_mem
module for convolutional volumes. It sends the data from the memory to the conv
op along with any padding required. It is able to generate all data streams for
the image segment in the memory including the extra needed for maxpools.