Add buffer to image_read to enable stalling of data
A FIFO is used to enable the stalling of the image data when the 'image_rdy'
flag is deasserted. The buffer is required to allow the pipelined data to drain
as the logic for generating the address and data being received from the
image_mem module and not easily be stalled.