Replace STR_IMG_WIDTH with IMG_WIDTH*DEPTH_NB
The optimal width of the data bus for bring data into the image_write and
image_mem modules is IMG_WIDTH*DEPTH_NB. This is because it is the largest
contiguous and addressable section of data that can be written to the
image_mem. Any larger and it would have to be broken up for the scatter
operation that is writing the data to image_mem. Any less and we would have to
generate multiple address for writing data in sequence.