Add scripts to generate bitstream
The task of building the workspace and project directories has been contained
in one script while the work of calling steps to generate the bitstream file is
in another. This allows for a developer to more more easily swap between the
Vivado GUI or batch mode.

Both scripts are based on those from 'https://github.com/bmartini/zynq-axis'.