Simplify fifo_simple 'count' calculation
By using the pointers instead of the address for the calculation we can
simplify the expression while being functionally and mathematically the same.

Also of note is that the calculation is using the "_nx" next values of the
pointers instead of just one so that the count will now be an exact
representation of the number of values stored in the FIFO. This may give
different zero and max counts to the indicated empty and full signals due to
the conservative triggering logic of the empty/full signals.