Update fifo_simple formal module for robustness
First is changing how we set the initial reset state using 'restrict' instead
of 'assume'.  The restrict acts just like the assume during formal verification
but will be ignored in simulators if we wanted to also use the asserts during
simulation.

Add an impossible assert, since the FIFO uses comparisons of the same
counter/pointers to calculate the empty/full flags, if there was a buggy bit of
logic this assert would find it.

Make the assert being tested with the rise of the full signal more robust.
Since we are using the 'past' function the logic needs to be sure it does not
test any lingering state from before a reset.