Move fifo_simple to strict behavior for empty/full
Before the empty and full signals were slow to deassert. This was done due to
trying to have more registered values being used in the calculation of the
signals. However, testing showed that the new way of generating the signals is
able to be synth to a smaller circuit. It also makes the count and empty/full
signals trigger in lock step.

The formal module is also simplified.