Limit pool cnt assert to only test valid configurations
If the layer module is sent a configuration after the maxpool has started being
processed, the configuration and thus the module is in a 'invalid'
configuration state. This change ensure that the assert is checked only if the
module has been configured before the first maxpool value has been calculated.