Turn off for YOSYS restoring `default_nettype wire
The Yosys program cannot selectively turn on and off the default_nettype for
individual modules and will only pay attention to the last value set. This is
not the case for other tools and it is good practice the restore the
default_nettype when mixing these modules this 3rd party modules.