Add pybind11 interface to Verilator simulation
Only two functions need to be implemented to customize the interface for a
module. The 'prep' function needs to set any input of the dut. The 'update'
function takes all inputs and outputs of the dut add wraps it into a python
dict to be used by the python testbench.